+++ Source/Device.generated.swift
@@ -66,7 +66,7 @@ public enum Device {
     case iPhone4
     /// Device is an [iPhone 4s](https://support.apple.com/kb/SP643)
     ///
-    /// ![Image](https://support.apple.com/library/content/dam/edam/applecare/images/en_US/iphone/iphone5s/iphone_4s.png)
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP643/sp643_iphone4s_color_black.jpg)
     case iPhone4s
     /// Device is an [iPhone 5](https://support.apple.com/kb/SP655)
     ///
@@ -132,6 +132,18 @@ public enum Device {
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP781/SP781-iPhone-xr.jpg)
     case iPhoneXR
+    /// Device is an [iPhone 11](https://support.apple.com/kb/SP804)
+    ///
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP804/sp804-iphone11_2x.png)
+    case iPhone11
+    /// Device is an [iPhone 11 Pro](https://support.apple.com/kb/SP805)
+    ///
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP805/sp805-iphone11pro_2x.png)
+    case iPhone11Pro
+    /// Device is an [iPhone 11 Pro Max](https://support.apple.com/kb/SP806)
+    ///
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP806/sp806-iphone11pro-max_2x.png)
+    case iPhone11ProMax
     /// Device is an [iPad 2](https://support.apple.com/kb/SP622)
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP622/SP622_01-ipad2-mul.png)
@@ -158,7 +170,7 @@ public enum Device {
     case iPad5
     /// Device is an [iPad (6th generation)](https://support.apple.com/kb/SP774)
     ///
-    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP751/ipad_5th_generation.png)
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP774/sp774-ipad-6-gen_2x.png)
     case iPad6
     /// Device is an [iPad Air (3rd generation)](https://support.apple.com/kb/SP787)
     ///
@@ -262,6 +274,14 @@ public enum Device {
     ///
     /// ![Image](https://km.support.apple.com/resources/sites/APPLE/content/live/IMAGES/0/IM911/en_US/aw-series4-nike-240.png)
     case appleWatchSeries4_44mm
+    /// Device is an [Apple Watch Series 5](https://support.apple.com/kb/SP808)
+    ///
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP808/sp808-apple-watch-series-5_2x.png)
+    case appleWatchSeries5_40mm
+    /// Device is an [Apple Watch Series 5](https://support.apple.com/kb/SP808)
+    ///
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP808/sp808-apple-watch-series-5_2x.png)
+    case appleWatchSeries5_44mm
   #endif
 
   /// Device is [Simulator](https://developer.apple.com/library/ios/documentation/IDEs/Conceptual/iOS_Simulator_Guide/Introduction/Introduction.html)
@@ -321,6 +341,9 @@ public enum Device {
       case "iPhone11,2": return iPhoneXS
       case "iPhone11,4", "iPhone11,6": return iPhoneXSMax
       case "iPhone11,8": return iPhoneXR
+      case "iPhone12,1": return iPhone11
+      case "iPhone12,3": return iPhone11Pro
+      case "iPhone12,5": return iPhone11ProMax
       case "iPad2,1", "iPad2,2", "iPad2,3", "iPad2,4": return iPad2
       case "iPad3,1", "iPad3,2", "iPad3,3": return iPad3
       case "iPad3,4", "iPad3,5", "iPad3,6": return iPad4
@@ -363,6 +386,8 @@ public enum Device {
       case "Watch3,2", "Watch3,4": return appleWatchSeries3_42mm
       case "Watch4,1", "Watch4,3": return appleWatchSeries4_40mm
       case "Watch4,2", "Watch4,4": return appleWatchSeries4_44mm
+      case "Watch5,1", "Watch5,3": return appleWatchSeries5_40mm
+      case "Watch5,2", "Watch5,4": return appleWatchSeries5_44mm
       case "i386", "x86_64": return simulator(mapToDevice(identifier: ProcessInfo().environment["SIMULATOR_MODEL_IDENTIFIER"] ?? "watchOS"))
       default: return unknown(identifier)
       }
@@ -410,6 +435,9 @@ public enum Device {
         case .iPhoneXS: return 5.8
         case .iPhoneXSMax: return 6.5
         case .iPhoneXR: return 6.1
+        case .iPhone11: return 6.1
+        case .iPhone11Pro: return 5.8
+        case .iPhone11ProMax: return 6.5
         case .iPad2: return 9.7
         case .iPad3: return 9.7
         case .iPad4: return 9.7
@@ -445,6 +473,8 @@ public enum Device {
       case .appleWatchSeries3_42mm: return 1.6
       case .appleWatchSeries4_40mm: return 1.8
       case .appleWatchSeries4_44mm: return 2.0
+      case .appleWatchSeries5_40mm: return 1.8
+      case .appleWatchSeries5_44mm: return 2.0
       case .simulator(let model): return model.diagonal
       case .unknown: return -1
       }
@@ -477,6 +507,9 @@ public enum Device {
       case .iPhoneXS: return (width: 9, height: 19.5)
       case .iPhoneXSMax: return (width: 9, height: 19.5)
       case .iPhoneXR: return (width: 9, height: 19.5)
+      case .iPhone11: return (width: 9, height: 19.5)
+      case .iPhone11Pro: return (width: 9, height: 19.5)
+      case .iPhone11ProMax: return (width: 9, height: 19.5)
       case .iPad2: return (width: 3, height: 4)
       case .iPad3: return (width: 3, height: 4)
       case .iPad4: return (width: 3, height: 4)
@@ -512,6 +545,8 @@ public enum Device {
       case .appleWatchSeries3_42mm: return (width: 4, height: 5)
       case .appleWatchSeries4_40mm: return (width: 4, height: 5)
       case .appleWatchSeries4_44mm: return (width: 4, height: 5)
+      case .appleWatchSeries5_40mm: return (width: 4, height: 5)
+      case .appleWatchSeries5_44mm: return (width: 4, height: 5)
       case .simulator(let model): return model.screenRatio
       case .unknown: return (width: -1, height: -1)
       }
@@ -528,7 +563,7 @@ public enum Device {
 
     /// All iPhones
     public static var allPhones: [Device] {
-      return [.iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR]
+      return [.iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax]
     }
 
     /// All iPads
@@ -538,17 +573,17 @@ public enum Device {
 
     /// All X-Series Devices
     public static var allXSeriesDevices: [Device] {
-      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR]
+      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax]
     }
 
     /// All Plus-Sized Devices
     public static var allPlusSizedDevices: [Device] {
-      return [.iPhone6Plus, .iPhone6sPlus, .iPhone7Plus, .iPhone8Plus]
+      return [.iPhone6Plus, .iPhone6sPlus, .iPhone7Plus, .iPhone8Plus, .iPhoneXSMax, .iPhone11ProMax]
     }
 
     /// All Pro Devices
     public static var allProDevices: [Device] {
-      return [.iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3]
+      return [.iPhone11Pro, .iPhone11ProMax, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3]
     }
 
     /// All mini Devices
@@ -639,7 +674,7 @@ public enum Device {
 
     /// All Face ID Capable Devices
     public static var allFaceIDCapableDevices: [Device] {
-      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPadPro11Inch, .iPadPro12Inch3]
+      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPadPro11Inch, .iPadPro12Inch3]
     }
 
     /// Returns whether or not the device has Touch ID
@@ -659,7 +694,7 @@ public enum Device {
 
     /// All devices that feature a sensor housing in the screen
     public static var allDevicesWithSensorHousing: [Device] {
-      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR]
+      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax]
     }
 
     /// Returns whether or not the device has a sensor housing
@@ -669,7 +704,7 @@ public enum Device {
 
     /// All devices that feature a screen with rounded corners.
     public static var allDevicesWithRoundedDisplayCorners: [Device] {
-      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPadPro11Inch, .iPadPro12Inch3]
+      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPadPro11Inch, .iPadPro12Inch3]
     }
 
     /// Returns whether or not the device has a screen with rounded corners.
@@ -699,7 +734,7 @@ public enum Device {
   #elseif os(watchOS)
     /// All Watches
     public static var allWatches: [Device] {
-       return [.appleWatchSeries0_38mm, .appleWatchSeries0_42mm, .appleWatchSeries1_38mm, .appleWatchSeries1_42mm, .appleWatchSeries2_38mm, .appleWatchSeries2_42mm, .appleWatchSeries3_38mm, .appleWatchSeries3_42mm, .appleWatchSeries4_40mm, .appleWatchSeries4_44mm]
+       return [.appleWatchSeries0_38mm, .appleWatchSeries0_42mm, .appleWatchSeries1_38mm, .appleWatchSeries1_42mm, .appleWatchSeries2_38mm, .appleWatchSeries2_42mm, .appleWatchSeries3_38mm, .appleWatchSeries3_42mm, .appleWatchSeries4_40mm, .appleWatchSeries4_44mm, .appleWatchSeries5_40mm, .appleWatchSeries5_44mm]
     }
 
     /// All simulator Watches
@@ -709,7 +744,7 @@ public enum Device {
 
     /// All watches that have Force Touch support.
     public static var allWatchesWithForceTouchSupport: [Device] {
-      return [.appleWatchSeries0_38mm, .appleWatchSeries0_42mm, .appleWatchSeries1_38mm, .appleWatchSeries1_42mm, .appleWatchSeries2_38mm, .appleWatchSeries2_42mm, .appleWatchSeries3_38mm, .appleWatchSeries3_42mm, .appleWatchSeries4_40mm, .appleWatchSeries4_44mm]
+      return [.appleWatchSeries0_38mm, .appleWatchSeries0_42mm, .appleWatchSeries1_38mm, .appleWatchSeries1_42mm, .appleWatchSeries2_38mm, .appleWatchSeries2_42mm, .appleWatchSeries3_38mm, .appleWatchSeries3_42mm, .appleWatchSeries4_40mm, .appleWatchSeries4_44mm, .appleWatchSeries5_40mm, .appleWatchSeries5_44mm]
     }
 
     /// Returns whether or not the device has Force Touch support.
@@ -847,6 +882,9 @@ public enum Device {
       case .iPhoneXS: return 458
       case .iPhoneXSMax: return 458
       case .iPhoneXR: return 326
+      case .iPhone11: return 326
+      case .iPhone11Pro: return 458
+      case .iPhone11ProMax: return 458
       case .iPad2: return 132
       case .iPad3: return 264
       case .iPad4: return 264
@@ -882,6 +920,8 @@ public enum Device {
     case .appleWatchSeries3_42mm: return 303
     case .appleWatchSeries4_40mm: return 326
     case .appleWatchSeries4_44mm: return 326
+    case .appleWatchSeries5_40mm: return 326
+    case .appleWatchSeries5_44mm: return 326
     case .simulator(let model): return model.ppi
     case .unknown: return nil
     }
@@ -941,6 +981,9 @@ extension Device: CustomStringConvertible {
       case .iPhoneXS: return "iPhone Xs"
       case .iPhoneXSMax: return "iPhone Xs Max"
       case .iPhoneXR: return "iPhone Xʀ"
+      case .iPhone11: return "iPhone 11"
+      case .iPhone11Pro: return "iPhone 11 Pro"
+      case .iPhone11ProMax: return "iPhone 11 Pro Max"
       case .iPad2: return "iPad 2"
       case .iPad3: return "iPad (3rd generation)"
       case .iPad4: return "iPad (4th generation)"
@@ -976,6 +1019,8 @@ extension Device: CustomStringConvertible {
       case .appleWatchSeries3_42mm: return "Apple Watch Series 3 42mm"
       case .appleWatchSeries4_40mm: return "Apple Watch Series 4 40mm"
       case .appleWatchSeries4_44mm: return "Apple Watch Series 4 44mm"
+      case .appleWatchSeries5_40mm: return "Apple Watch Series 5 40mm"
+      case .appleWatchSeries5_44mm: return "Apple Watch Series 5 44mm"
       case .simulator(let model): return "Simulator (\(model))"
       case .unknown(let identifier): return identifier
       }
@@ -1269,6 +1314,7 @@ extension Device {
   public enum CameraTypes {
     case normal
     case telephoto
+    case ultraWide
   }
 
   /// Returns an array of the types of cameras the device has
@@ -1314,23 +1360,31 @@ extension Device {
       case .iPhoneX: return [.normal, .telephoto]
       case .iPhoneXS: return [.normal, .telephoto]
       case .iPhoneXSMax: return [.normal, .telephoto]
+      case .iPhone11: return [.normal, .ultraWide]
+      case .iPhone11Pro: return [.normal, .telephoto, .ultraWide]
+      case .iPhone11ProMax: return [.normal, .telephoto, .ultraWide]
       default: return []
     }
   }
 
   /// All devices that feature a camera
   public static var allDevicesWithCamera: [Device] {
-    return [.iPodTouch5, .iPodTouch6, .iPodTouch7, .iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3]
+    return [.iPodTouch5, .iPodTouch6, .iPodTouch7, .iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3]
   }
 
   /// All devices that feature a normal camera
   public static var allDevicesWithNormalCamera: [Device] {
-    return [.iPodTouch5, .iPodTouch6, .iPodTouch7, .iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3]
+    return [.iPodTouch5, .iPodTouch6, .iPodTouch7, .iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3]
   }
 
   /// All devices that feature a telephoto camera
   public static var allDevicesWithTelephotoCamera: [Device] {
-    return [.iPhone7Plus, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax]
+    return [.iPhone7Plus, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhone11Pro, .iPhone11ProMax]
+  }
+
+  /// All devices that feature an ultra wide camera
+  public static var allDevicesWithUltraWideCamera: [Device] {
+    return [.iPhone11, .iPhone11Pro, .iPhone11ProMax]
   }
 
   /// Returns whether or not the current device has a camera
@@ -1346,6 +1400,11 @@ extension Device {
   /// Returns whether or not the current device has a telephoto camera
   public var hasTelephotoCamera: Bool {
     return self.cameras.contains(.telephoto)
+  }
+
+  /// Returns whether or not the current device has an ultra wide camera
+  public var hasUltraWideCamera: Bool {
+    return self.cameras.contains(.ultraWide)
   }
 
 }